+++ ast-opt/propagate/opt_propagate.c
@@ -6,7 +6,7 @@
 
 #define MAX_CONSTANTS 32
 
-typedef struct { String name; long value; int is_unsigned; int is_long; int active; } ConstEntry;
+typedef struct { String name; long long value; int is_unsigned; int is_long; int active; } ConstEntry;
 
 /* from sub-files */
 extern void scan_node(AST_Node* n, ConstEntry* map, int* count);
@@ -35,7 +35,7 @@ ConstEntry* find_entry(ConstEntry* map, int count, String name)
     return NULL;
 }
 
-void add_entry(ConstEntry* map, int* count, String name, long value,
+void add_entry(ConstEntry* map, int* count, String name, long long value,
                int is_unsigned, int is_long)
 {
     if (*count >= MAX_CONSTANTS) return;
